add sign count

+++ get_util.php
@@ -2,7 +2,7 @@
 
     function connet_mysql()
     {
-        $con = mysql_connect("localhost", "root", "root");
+        $con = mysql_connect("127.0.0.1", "root", "root");
         if (! $con) {
             die('Could not connect: ' . mysql_error()) . "<br>";
         } 
@@ -17,7 +17,10 @@
     {
     
         $con = connet_mysql();
-        #mysql_query("set names 'utf8'");
+        if(PHP_OS == Darwin)
+        {
+        mysql_query("set names 'utf8'");
+        }
         mysql_select_db("ipsearch_db", $con);
         $select = "SELECT * FROM ipaddress";
         $result = mysql_query($select);
@@ -29,7 +32,10 @@
     {
         
         $con = connet_mysql();
-        #mysql_query("set names 'utf8'");
+        if(PHP_OS == Darwin)
+        {
+        mysql_query("set names 'utf8'");
+        }
         mysql_select_db("ipsearch_db", $con);
         $select = "SELECT * FROM ipaddress WHERE ip='" . $ip_address . "'";
         $result = mysql_query($select);
@@ -42,7 +48,10 @@
         
         $localtime_assoc = localtime(time(), true); 
         $con = connet_mysql();
-        #mysql_query("set names 'utf8'");
+        if(PHP_OS == Darwin)
+        {
+        mysql_query("set names 'utf8'");
+        }
         mysql_select_db("ipsearch_db", $con);
         $insert = "INSERT INTO ipaddress (ip,message,recordtime) VALUES ('$ip_address','$ip_message',now())";
         $result = mysql_query($insert);
@@ -54,7 +63,10 @@
     {
         $localtime_assoc = localtime(time(), true);
         $con = connet_mysql();
-        #mysql_query("set names 'utf8'");
+        if(PHP_OS == Darwin)
+        {
+        mysql_query("set names 'utf8'");
+        }
         mysql_select_db("ipsearch_db", $con);
         $insert = "UPDATE ipaddress  SET message='$ip_message',updatetime = now() where ip='$ip_address'";
         $result = mysql_query($insert);
@@ -65,7 +77,10 @@
     function del_ip_message($ip_address)
     {
         $con = connet_mysql();
-        #mysql_query("set names 'utf8'");
+    if(PHP_OS == Darwin)
+        {
+        mysql_query("set names 'utf8'");
+        }
         mysql_select_db("ipsearch_db", $con);
         $delect = "DELETE FROM ipaddress WHERE ip='$ip_address'";
         $result = mysql_query($delect);
@@ -76,7 +91,10 @@
     function get_ip_count()
     {
         $con = connet_mysql();
-        #mysql_query("set names 'utf8'");
+        if(PHP_OS == Darwin)
+        {
+        mysql_query("set names 'utf8'");
+        }
         mysql_select_db("ipsearch_db", $con);
         $count = "SELECT COUNT(*) FROM ipaddress";
         $result = mysql_query($count);
